Fixed Codacy complexity issue
