+++ test_modules/test_actions.py
@@ -1,7 +1,6 @@
 import unittest
 import mesh_model.mesh_struct.mesh as mesh
 from mesh_model.mesh_struct.mesh_elements import Dart, Node
-import numpy.testing
 from plots.mesh_plotter import plot_mesh
 
 from actions.triangular_actions import split_edge, flip_edge, collapse_edge
